allows a job's action to take in the subscriber and an optional worker. Remove empty_job factory function.
This allows an async job to schedules itself on the same worker each time, this pins the job to the same thread for each execution, on my 3900X it shows performance improvement of ~30%.

The previous "hack" for async job is flawed as it uses a dummy empty job to schedule itself to the back of the queue. While this works, each invocation of a job will use a different thread from the threadpool, this causes a job's work to shuffle around multiple threads, possibly triggering lots of unneeded cache movements.