Implement Ingestors
Split the existing Dispensers into Dispensers and Ingestors. Ingestors
will handle the taking in of objects delivered to them.

Added a new set of messages for the Ingestors, which mimics the existing
Dispenser messages.

Modified the Delivery task to make use of the Ingestors.

Added unit tests for the Ingestor and modified the existing Delivery test
to use a flaky Ingestor instead of a flaky Dispenser.